Move user interfaces to new types file in serve folder
