--- src/server/controllers/userControllers/types.ts
@@ -1,8 +1,0 @@
-import type { JwtPayload } from "jsonwebtoken";
-import type { UserStructure } from "../../../types/types";
-
-export interface CustomTokenPayload
-  extends Pick<UserStructure, "name" | "isAdmin">,
-    JwtPayload {
-  id: string;
-}
