+++ src/types/types.ts
@@ -1,29 +1,3 @@
-export interface UserCredentials {
-  email: string;
-  password: string;
-}
-
-export interface UserActivationCredentials
-  extends Pick<UserCredentials, "password"> {
-  confirmPassword: string;
-}
-
-export interface UserData extends Omit<UserCredentials, "password"> {
-  name: string;
-}
-
-export interface UserStructure
-  extends UserData,
-    Pick<UserCredentials, "password"> {
-  isAdmin: boolean;
-  isActive: boolean;
-  activationKey: string;
-}
-
-export interface UserWithId extends UserStructure {
-  _id: string;
-}
-
 export type WithRequiredProperties<T, K extends keyof T> = T & {
   [P in K]-?: T[P];
 };
